took actix out of mod names and started refactor
